requirements.txt
python-dotenv
flask
requests
streamlit
beautifulsoup4
langchain_community
langchain_huggingface
langchain_groq
faiss-cpu
